requirements.txt
requests==2.28.2
terminaltables==3.1.10
python-dotenv==1.0.0